requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
click>=8.1.0
pytest>=7.4.0
fastapi>=0.110.0
uvicorn>=0.27.0
pyjwt>=2.8.0
bcrypt>=4.0.0
google-auth>=2.20.0
pymupdf>=1.24.0
python-multipart>=0.0.9
apscheduler>=3.10.0
redis>=5.0.0
celery>=5.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0
